requirements.txt
langchain
langchain-community
langchain-astradb
langchain-huggingface
langchain-groq
datasets
pypdf
python-dotenv
pandas
flask
prometheus_client
